pyproject.toml
[project]
name = "copy-checker-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
